LabAPI from NuGet + Default DoorPermissionCheck + LabAPI wrapper overloads for permission checks
Changed the LabAPI references to be obtained via NuGet.
Added a `Default` value to `DoorPermissionCheck` that reflects base-game behavior.
Added overloads for checking permissions of `Door` `LockerChamber` and `Generator` wrappers.
Removed a redundant null check.
Fixed the phrasing of a sentence.
Set the default value of the `checkFlags` parameter to the new `Default` enum value.